+++ internal/diagnostics/server.go
@@ -2,26 +2,45 @@ package diagnostics
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"net/http/pprof"
+	"sync"
 
 	"github.com/go-logr/logr"
-	"github.com/kong/kubernetes-ingress-controller/internal/manager"
+	"github.com/kong/deck/file"
+
+	"github.com/kong/kubernetes-ingress-controller/internal/util"
 )
 
-// Server is an HTTP server running exposing the pprof profiling tool.
+// Server is an HTTP server running exposing the pprof profiling tool, and processing diagnostic dumps of Kong configurations.
 type Server struct {
-	Logger logr.Logger
+	Logger           logr.Logger
+	ProfilingEnabled bool
+	ConfigDumps      util.ConfigDumpDiagnostic
+	ConfigLock       *sync.RWMutex
 }
 
-// Listen starts up the HTTP server and blocks until ctx expires.
-func (s *Server) Listen(ctx context.Context) error {
-	mux := http.NewServeMux()
-	installHandlers(mux)
+var successfulConfigDump file.Content
+var failedConfigDump file.Content
 
-	httpServer := &http.Server{Addr: fmt.Sprintf(":%d", manager.DiagnosticsPort), Handler: mux}
+// Listen starts up the HTTP server and blocks until ctx expires.
+func (s *Server) Listen(ctx context.Context, port int) error {
+
+	mux := http.NewServeMux()
+	if s.ConfigDumps != (util.ConfigDumpDiagnostic{}) {
+		s.installDumpHandlers(mux)
+	}
+	if s.ProfilingEnabled {
+		installProfilingHandlers(mux)
+	}
+
+	httpServer := &http.Server{Addr: fmt.Sprintf(":%d", port), Handler: mux}
 	errChan := make(chan error)
+
+	go s.receiveConfig(ctx)
+
 	go func() {
 		err := httpServer.ListenAndServe()
 		if err != nil {
@@ -35,7 +54,7 @@ func (s *Server) Listen(ctx context.Context) error {
 		}
 	}()
 
-	s.Logger.Info("diagnostics server is starting to listen", "addr", manager.DiagnosticsPort)
+	s.Logger.Info("diagnostics server is starting to listen", "addr", port)
 
 	select {
 	case <-ctx.Done():
@@ -46,8 +65,30 @@ func (s *Server) Listen(ctx context.Context) error {
 	}
 }
 
-// installHandlers adds the Profiling webservice to the given mux.
-func installHandlers(mux *http.ServeMux) {
+// receiveConfig watches the config update channel
+func (s *Server) receiveConfig(ctx context.Context) {
+	for {
+		select {
+		case dump := <-s.ConfigDumps.Configs:
+			s.ConfigLock.Lock()
+			if dump.Failed {
+				failedConfigDump = dump.Config
+			} else {
+				successfulConfigDump = dump.Config
+			}
+			s.ConfigLock.Unlock()
+		case <-ctx.Done():
+			if err := ctx.Err(); err != nil {
+				s.Logger.Error(err, "shutting down diagnostic config collection: context completed with error")
+			}
+			s.Logger.V(3).Info("shutting down diagnostic config collection: context completed")
+			return
+		}
+	}
+}
+
+// installProfilingHandlers adds the Profiling webservice to the given mux.
+func installProfilingHandlers(mux *http.ServeMux) {
 	mux.HandleFunc("/debug/pprof", redirectTo("/debug/pprof/"))
 	mux.HandleFunc("/debug/pprof/", pprof.Index)
 	mux.HandleFunc("/debug/pprof/heap", pprof.Index)
@@ -61,9 +102,26 @@ func installHandlers(mux *http.ServeMux) {
 	mux.HandleFunc("/debug/pprof/trace", pprof.Trace)
 }
 
+// installDumpHandlers adds the config dump webservice to the given mux.
+func (s *Server) installDumpHandlers(mux *http.ServeMux) {
+	mux.HandleFunc("/debug/config/successful", s.lastConfig(&successfulConfigDump))
+	mux.HandleFunc("/debug/config/failed", s.lastConfig(&failedConfigDump))
+}
+
 // redirectTo redirects request to a certain destination.
 func redirectTo(to string) func(http.ResponseWriter, *http.Request) {
 	return func(rw http.ResponseWriter, req *http.Request) {
 		http.Redirect(rw, req, to, http.StatusFound)
+	}
+}
+
+func (s *Server) lastConfig(config *file.Content) func(rw http.ResponseWriter, req *http.Request) {
+	return func(rw http.ResponseWriter, req *http.Request) {
+		rw.Header().Set("Content-Type", "application/json")
+		s.ConfigLock.RLock()
+		if err := json.NewEncoder(rw).Encode(*config); err != nil {
+			rw.WriteHeader(http.StatusInternalServerError)
+		}
+		s.ConfigLock.RUnlock()
 	}
 }