+++ internal/sendconfig/kong.go
@@ -4,6 +4,7 @@ import (
 	"github.com/blang/semver/v4"
 	"github.com/kong/deck/file"
 	"github.com/kong/go-kong/kong"
+
 	"github.com/kong/kubernetes-ingress-controller/internal/util"
 )