+++ internal/store/store.go
@@ -22,6 +22,7 @@ import (
 	"reflect"
 	"sort"
 	"strings"
+	"sync"
 
 	"github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
@@ -42,7 +43,6 @@ import (
 
 	"github.com/kong/kubernetes-ingress-controller/internal/annotations"
 	kongv1 "github.com/kong/kubernetes-ingress-controller/pkg/apis/configuration/v1"
-	"github.com/kong/kubernetes-ingress-controller/pkg/apis/configuration/v1beta1"
 	kongv1beta1 "github.com/kong/kubernetes-ingress-controller/pkg/apis/configuration/v1beta1"
 )
 
@@ -123,6 +123,8 @@ type CacheStores struct {
 	KongIngress   cache.Store
 
 	KnativeIngress cache.Store
+
+	l *sync.RWMutex
 }
 
 // NewCacheStores is a convenience function for CacheStores to initialize all attributes with new cache stores
@@ -139,6 +141,7 @@ func NewCacheStores() (c CacheStores) {
 	c.TCPIngress = cache.NewStore(keyFunc)
 	c.UDPIngress = cache.NewStore(keyFunc)
 	c.KongIngress = cache.NewStore(keyFunc)
+	c.l = &sync.RWMutex{}
 	return
 }
 
@@ -186,6 +189,9 @@ func NewCacheStoresFromObjs(objs ...runtime.Object) (CacheStores, error) {
 
 // Get checks whether or not there's already some version of the provided object present in the cache.
 func (c CacheStores) Get(obj runtime.Object) (item interface{}, exists bool, err error) {
+	c.l.RLock()
+	defer c.l.RUnlock()
+
 	switch obj := obj.(type) {
 	// ----------------------------------------------------------------------------
 	// Kubernetes Core API Support
@@ -227,6 +233,9 @@ func (c CacheStores) Get(obj runtime.Object) (item interface{}, exists bool, err
 // Add stores a provided runtime.Object into the CacheStore if it's of a supported type.
 // The CacheStore must be initialized (see NewCacheStores()) or this will panic.
 func (c CacheStores) Add(obj runtime.Object) error {
+	c.l.Lock()
+	defer c.l.Unlock()
+
 	switch obj := obj.(type) {
 	// ----------------------------------------------------------------------------
 	// Kubernetes Core API Support
@@ -269,6 +278,9 @@ func (c CacheStores) Add(obj runtime.Object) error {
 // Delete removes a provided runtime.Object from the CacheStore if it's of a supported type.
 // The CacheStore must be initialized (see NewCacheStores()) or this will panic.
 func (c CacheStores) Delete(obj runtime.Object) error {
+	c.l.Lock()
+	defer c.l.Unlock()
+
 	switch obj := obj.(type) {
 	// ----------------------------------------------------------------------------
 	// Kubernetes Core API Support
@@ -437,8 +449,8 @@ func (s Store) ListTCPIngresses() ([]*kongv1beta1.TCPIngress, error) {
 }
 
 // ListUDPIngresses returns the list of UDP Ingresses
-func (s Store) ListUDPIngresses() ([]*v1beta1.UDPIngress, error) {
-	ingresses := []*v1beta1.UDPIngress{}
+func (s Store) ListUDPIngresses() ([]*kongv1beta1.UDPIngress, error) {
+	ingresses := []*kongv1beta1.UDPIngress{}
 	if s.stores.UDPIngress == nil {
 		// older versions of the KIC do not support UDPIngress so short circuit to maintain support with them
 		return ingresses, nil
@@ -446,7 +458,7 @@ func (s Store) ListUDPIngresses() ([]*v1beta1.UDPIngress, error) {
 
 	err := cache.ListAll(s.stores.UDPIngress, labels.NewSelector(),
 		func(ob interface{}) {
-			ing, ok := ob.(*v1beta1.UDPIngress)
+			ing, ok := ob.(*kongv1beta1.UDPIngress)
 			if ok && s.isValidIngressClass(&ing.ObjectMeta, annotations.ExactClassMatch) {
 				ingresses = append(ingresses, ing)
 			}